# app/docs/api/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "GenAI Demo - DDD 電商平台 API",
    "description" : "基於領域驅動設計（DDD）的電商平台 API\n\n## 架構特色\n- 遵循 DDD 分層架構\n- 六邊形架構（Hexagonal Architecture）\n- CQRS 和事件驅動設計\n- 豐富的領域模型\n\n## 主要功能模組\n- 訂單管理：完整的訂單生命週期\n- 支付處理：多種支付方式和狀態管理\n- 庫存管理：即時庫存追蹤和預留\n- 產品管理：產品資訊和分類\n- 促銷活動：多樣化的促銷規則\n- 定價策略：動態定價和佣金計算\n- 客戶管理：會員等級和偏好設定\n\n## 錯誤處理\n所有 API 都遵循統一的錯誤回應格式，包含錯誤代碼、訊息、時間戳記和詳細資訊。\n",
    "contact" : {
      "name" : "GenAI Demo Team",
      "url" : "https://github.com/humank/genai-demo",
      "email" : "contact@genai-demo.com"
    },
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://test-api.genai-demo.com",
    "description" : "測試環境"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/payments" : {
      "get" : {
        "tags" : [ "支付管理" ],
        "summary" : "獲取所有支付記錄",
        "description" : "獲取系統中所有支付記錄的列表，通常用於管理後台查詢",
        "operationId" : "getAllPayments",
        "responses" : {
          "500" : {
            "description" : "系統內部錯誤",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "成功獲取支付列表",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "支付管理" ],
        "summary" : "處理支付",
        "description" : "處理新的支付請求，支援多種支付方式包括信用卡、轉帳等",
        "operationId" : "processPayment",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PaymentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "支付處理成功",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "422" : {
            "description" : "支付處理失敗（業務邏輯錯誤）",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "請求參數無效",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "系統內部錯誤",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments/{paymentId}/refund" : {
      "post" : {
        "tags" : [ "支付管理" ],
        "summary" : "處理退款",
        "description" : "對指定的支付進行退款處理，將款項退回給客戶",
        "operationId" : "refundPayment",
        "parameters" : [ {
          "name" : "paymentId",
          "in" : "path",
          "description" : "支付ID（UUID格式）",
          "required" : true,
          "schema" : {
            "pattern" : "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
            "type" : "string"
          },
          "example" : "123e4567-e89b-12d3-a456-426614174000"
        } ],
        "responses" : {
          "422" : {
            "description" : "支付狀態不允許退款",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "支付ID格式無效",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "系統內部錯誤",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "退款處理成功",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "支付不存在",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments/{paymentId}/cancel" : {
      "post" : {
        "tags" : [ "支付管理" ],
        "summary" : "取消支付",
        "description" : "取消指定的支付，通常用於處理中的支付或預授權支付",
        "operationId" : "cancelPayment",
        "parameters" : [ {
          "name" : "paymentId",
          "in" : "path",
          "description" : "支付ID（UUID格式）",
          "required" : true,
          "schema" : {
            "pattern" : "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
            "type" : "string"
          },
          "example" : "123e4567-e89b-12d3-a456-426614174000"
        } ],
        "responses" : {
          "400" : {
            "description" : "支付ID格式無效",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "422" : {
            "description" : "支付狀態不允許取消",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "系統內部錯誤",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "支付取消成功",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "支付不存在",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/payments/{paymentId}" : {
      "get" : {
        "tags" : [ "支付管理" ],
        "summary" : "獲取支付詳情",
        "description" : "根據支付ID獲取特定支付的詳細資訊",
        "operationId" : "getPayment",
        "parameters" : [ {
          "name" : "paymentId",
          "in" : "path",
          "description" : "支付ID（UUID格式）",
          "required" : true,
          "schema" : {
            "pattern" : "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
            "type" : "string"
          },
          "example" : "123e4567-e89b-12d3-a456-426614174000"
        } ],
        "responses" : {
          "400" : {
            "description" : "支付ID格式無效",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "系統內部錯誤",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "成功獲取支付詳情",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "支付不存在",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments/order/{orderId}" : {
      "get" : {
        "tags" : [ "支付管理" ],
        "summary" : "根據訂單ID獲取支付",
        "description" : "根據訂單ID獲取該訂單相關的支付資訊",
        "operationId" : "getPaymentByOrderId",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "description" : "訂單ID（UUID格式）",
          "required" : true,
          "schema" : {
            "pattern" : "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
            "type" : "string"
          },
          "example" : "123e4567-e89b-12d3-a456-426614174000"
        } ],
        "responses" : {
          "200" : {
            "description" : "成功獲取訂單支付詳情",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "訂單ID格式無效",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "訂單支付不存在",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "系統內部錯誤",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "StandardErrorResponse" : {
        "required" : [ "code", "message", "path", "timestamp" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "錯誤代碼",
            "example" : "VALIDATION_ERROR"
          },
          "message" : {
            "type" : "string",
            "description" : "錯誤訊息",
            "example" : "請求參數驗證失敗"
          },
          "timestamp" : {
            "type" : "string",
            "description" : "錯誤發生時間",
            "format" : "date-time"
          },
          "path" : {
            "type" : "string",
            "description" : "請求路徑",
            "example" : "/api/orders"
          },
          "details" : {
            "type" : "array",
            "description" : "詳細錯誤資訊",
            "nullable" : true,
            "items" : {
              "$ref" : "#/components/schemas/FieldError"
            }
          },
          "traceId" : {
            "type" : "string",
            "description" : "追蹤ID，用於問題排查",
            "nullable" : true,
            "example" : "abc123-def456-ghi789"
          }
        },
        "description" : "標準錯誤回應"
      },
      "FieldError" : {
        "required" : [ "field", "message" ],
        "type" : "object",
        "properties" : {
          "field" : {
            "type" : "string",
            "description" : "欄位名稱",
            "example" : "customerId"
          },
          "message" : {
            "type" : "string",
            "description" : "錯誤訊息",
            "example" : "客戶ID不能為空"
          },
          "rejectedValue" : {
            "type" : "object",
            "description" : "拒絕的值",
            "nullable" : true,
            "example" : null
          },
          "code" : {
            "type" : "string",
            "description" : "錯誤代碼",
            "nullable" : true,
            "example" : "NotBlank"
          }
        },
        "description" : "欄位錯誤詳情",
        "nullable" : true
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "請在此處輸入 JWT Token，格式：Bearer {token}",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "apiKey" : {
        "type" : "apiKey",
        "description" : "請在此處輸入 API Key",
        "name" : "X-API-Key",
        "in" : "header"
      }
    }
  }
}